Est par (#70)
* refactor: update noise reduction methods and enhance noise estimation functionality

* refactor: enhance SNR calculation and logging in spectrum processing

* refactor: add flat noise estimation method and update denoise method in spectrum processing

---------

+++ src/massflow/preprocess/numba/noise_reduction_numba.py
@@ -493,7 +493,7 @@ def smooth_signal_gaussian_numba(
 
     # Generate Gaussian kernel
     x = np.arange(-(window // 2), window // 2 + 1, dtype=np.float64)
-    kernel = np.exp(-0.5 * (x / sd) ** 2).astype(np.float64)
+    kernel = np.exp(-0.5 * (x / sd) ** 2)
     kernel /= kernel.sum()
 
     return convolve1d_flat_jit(intensity_arr, kernel, lengths_arr)
